Improve package names
