+++ src/main/scala/com/codacy/plugins/api/results/Parameter.scala
@@ -1,4 +1,4 @@
-package codacy.docker.api
+package com.codacy.plugins.api.results
 
 object Parameter {
